requirements.txt
requests==2.18.4
psycopg2==2.7.5
python-dateutil==2.7.5